chore(providers): Named exports and index.ts exports added
