+++ packages/core/src/providers/index.ts
@@ -1,0 +1,4 @@
+export * from "./email/index.js";
+export * from "./newsletter/index.js";
+export * from "./payment/index.js";
+export * from "./payment-flow/index.js";
