+++ packages/core/src/providers/newsletter/NoneProvider.ts
@@ -4,7 +4,7 @@ import {
   UpdateNewsletterContact
 } from "#type/index";
 
-export default class NoneProvider implements NewsletterProvider {
+export class NoneProvider implements NewsletterProvider {
   async addTagToContacts(emails: string[], tag: string): Promise<void> {}
   async removeTagFromContacts(emails: string[], tag: string): Promise<void> {}
   async getContact(email: string): Promise<NewsletterContact | undefined> {
@@ -27,3 +27,6 @@ export default class NoneProvider implements NewsletterProvider {
     fields: Record<string, string>
   ): Promise<void> {}
 }
+
+/** @deprecated Use named import NoneProvider instead */
+export default NoneProvider;
