+++ packages/core/src/providers/newsletter/index.ts
@@ -1,0 +1,2 @@
+export * from "./MailchimpProvider.js";
+export * from "./NoneProvider.js";
