+++ packages/core/src/providers/payment-flow/PaymentFlowProvider.ts
@@ -1,0 +1,23 @@
+import { JoinFlow } from "#models/index";
+import {
+  CompletedPaymentFlow,
+  CompletedPaymentFlowData,
+  PaymentFlow,
+  PaymentFlowData
+} from "#type/index";
+
+export abstract class PaymentFlowProvider {
+  abstract createPaymentFlow(
+    joinFlow: JoinFlow,
+    completeUrl: string,
+    data: PaymentFlowData
+  ): Promise<PaymentFlow>;
+
+  abstract completePaymentFlow(
+    joinFlow: JoinFlow
+  ): Promise<CompletedPaymentFlow>;
+
+  abstract getCompletedPaymentFlowData(
+    completedPaymentFlow: CompletedPaymentFlow
+  ): Promise<CompletedPaymentFlowData>;
+}
