+++ packages/core/src/type/index.ts
@@ -12,6 +12,7 @@ export * from "./contribution-info";
 export * from "./create-contact-mfa-data";
 export * from "./delete-contact-mfa-data";
 export * from "./email-mailing-recipient";
+export * from "./email";
 export * from "./export-type-id";
 export * from "./filter-handlers";
 export * from "./login-data";
